backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.20
trimesh==4.6.8
cairosvg==2.9.0
respx==0.23.1
pytest-asyncio==1.3.0
